% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindMean.R
\name{mean_of_each_group}
\alias{mean_of_each_group}
\title{mean_of_each_group function}
\usage{
mean_of_each_group(data, x, y)
}
\arguments{
\item{data}{is the dataset}

\item{x}{is the column that wanted to be grouped}

\item{y}{is the variable that we need to find the mean value}
}
\value{
the mean value of the grouped variable
}
\description{
This function can be used to output the mean value of each group
}
\examples{

Hotel_Dataset <- data.frame(
City = c("Vancouver", "Vancouver", "Vancouver", "Ames", "Ames", "Taipei"),
Location = c("Richmond", "Downtown", "Marine Drive", "Boulvard", "ISU", "Dehui"),
Vacancy_rooms = c(10, 5, 20, 2, 100, 50),
Price = c(200, 100, 250, 180, 200, 250)
)


}
